pyproject.toml
[project]
name = "project-1-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.50.0",
    "fpdf>=1.7.2",
    "hf-xet>=1.0.4",
    "httpx>=0.28.1",
    "huggingface-hub>=0.30.2",
    "num2words>=0.5.14",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "python-docx>=1.1.2",
    "pyttsx3>=2.98",
    "streamlit>=1.44.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]
